+++ itella-shipping/itella-shipping.php
@@ -13,9 +13,9 @@
  * @package           Itella_Shipping
  *
  * @wordpress-plugin
- * Plugin Name:       Smartpost Itella Shipping
+ * Plugin Name:       Smartposti Shipping
  * Plugin URI:
- * Description:       Smartpost Itella shipping plugin for Woocommerce
+ * Description:       Smartposti shipping plugin for Woocommerce
  * Version:           1.5.3
  * Author:            Itella Team
  * Author URI:        https://itella.lt/en/
@@ -24,8 +24,8 @@
  * Text Domain:       itella-shipping
  * Domain Path:       /languages
  *
- * Tested up to:      6.7.1
- * WC tested up to:   8.3.1
+ * Tested up to:      6.7.2
+ * WC tested up to:   9.7.0
  * Requires PHP:      7.2
  */