+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 ### Improved
 - added use of "Address 2" field value when registering a shipment
 
+### Changed
+- changed name "Smartpost" to "Smartposti"
+- changed logo
+- changed map colors
+
 ## [1.5.3] - 2025-01-16
 ### Improved
 - added to automatically recognize Woocommerce COD payment method